server/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
pillow==11.1.0
platformdirs==4.3.7
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.30.0
virtualenvwrapper-win==1.2.7
